requirements.txt
matplotlib
tensorflow-macos